base_global/lambda/handle-interaction/requirements.txt
boto3>=1.34.140
PyNaCl>=1.5.0
